+++ Src/HLGC/Prj1/source/ViewManager.cpp
@@ -285,7 +285,7 @@ void ViewManager::SetWindowSize(int width,int height) {
 	w_width = width;
 	w_height = height;
 	projMatrix = GetProjectionMatrix();
-	aspect = height / width;
+	aspect = (float)height / width;
 }
